add doctest for session extractor
